Refactor and improve code
Refactored code for readability and improved code logic when
converting image to ascii by using finding the luminance using a
specific but general formula instead of my own generic formula which
is very ineffecient and inaccurate.

Also, instead of printing the result/output on the terminal as bunch
of strings, a function will use Pillow to create a new raw image and
will draw the ascii character on it then save it in image folder.
It will then be displayed.