Return both raw image and converted when converting
Mainly, we restructure our code logic. We made changes to the models
having RawImage model and AsciiImage model. When the user uploads
the image we save the raw image and create a new converted image
then save it as AsciiImage instance. After conversion, the server
will both return the raw image and the converted image to the client
for rendering so that the user can see both.

Important note, everytime a user uploads and converts image, all
image instance and files will be automatically to maintain
lightweight storage capacity. The only data that will exist is the
users current data.